1. add early stop for lr & nn 2. change eval score from int to double to support raw score

+++ src/main/java/ml/shifu/shifu/core/posttrain/PostTrainMapper.java
@@ -208,7 +208,7 @@ public class PostTrainMapper extends Mapper<LongWritable, Text, IntWritable, Fea
         StringBuilder sb = new StringBuilder(500);
         sb.append(csr.getAvgScore()).append(Constants.DEFAULT_DELIMITER).append(csr.getMaxScore())
                 .append(Constants.DEFAULT_DELIMITER).append(csr.getMinScore()).append(Constants.DEFAULT_DELIMITER);
-        for(Integer score: csr.getScores()) {
+        for(Double score: csr.getScores()) {
             sb.append(score).append(Constants.DEFAULT_DELIMITER);
         }
         List<String> metaList = modelConfig.getMetaColumnNames();
